Added basic project skeleton
Main is the entry point for the program. It does the init and cleanup for SDL.
Application has a single method start which is where the interesting stuff stats.
Display is a wrapper for the SDL_Window and SDL_Renderer objects. Its job right now is to iterate through every GameObject in the level and display it.
GameObject is an abstract class which is supposed to represent every entity in the level. It has functionality for updating, handling events, and drawing of objects in the level.
Player is a demo implementation of AGameObject.
Vector is the vector class we did for homework. It is used to represent position.