Added functionality for picking up and putting down blocks
A pointer to the LevelState is passed to every GameObject as a way
of checking its surroundings.
The LeveState and GameObjects can now move around without occupying
the same space via a move method which checks if moves are valid.
The player can pick up blocks and put them down with the s key.